refactor: use main dir
